reorganise codebase
